rename list.js to lists
