+++ server/routes/index.js
@@ -1,7 +1,7 @@
 const searchRoute = require('./search');
 const loginRoute = require('./login');
 const logoutRoute = require('./logout');
-const listRoute = require('./list');
+const listRoute = require('./lists');
 
 module.exports = {
   searchRoute,
